rrr: convert rpc/request_options to module (rrr.request_options)
request_options.cpp is now the rrr.request_options module interface
unit (renamed from .hpp). Holds the RequestOptions POD, TimeoutType
enum, factory presets (defaults / with_retry / idempotent_retry /
no_timeout / fast / patient), and the jitter-based delay calculator.

GMF gotcha: `import std;` alone doesn't make `uint8_t` /
`uint16_t` / `uint64_t` available for type aliases used at struct-
member-declaration scope. Needs an explicit `#include <cstdint>` in
the GMF. (Inline function bodies that use `std::pow` etc. find them
fine through `import std;` — the issue is purely the declarative
typedefs.)

Includers (2) updated to `import rrr.request_options;`:
  - src/rrr/rpc/client.hpp, src/rrr/rrr.hpp.

Measurement (rrr + rpcbench, -j32, clean):
  +errors:           38.36 s, 175.0 MB PCM
  +request_options:  38.83 s, 175.2 MB PCM, librrr.a 9.71 MB

11 modules total; 2.30× baseline.
